Refresh

+++ src/apiCalls/removeFromFavoritesDb.js
@@ -1,14 +1,24 @@
+/* eslint-disable camelcase */
+import { getAccessToken } from "./getAccessToken";
+
 export const removeFromFavoritesDb = async (movieId, userId) => {
-  const favoriteData = { movieId, userId };
-  const url = `api/users/${userId}/favorites/${movieId}`;
+  const token = await getAccessToken();
+  if (!token) {
+    return null;
+  }
+  const favoriteData = { movie_id: movieId, user_id: userId };
+  const apiRoot = process.env.REACT_APP_AUTH_API_URL;
+  const url = apiRoot + "/users/favorites";
   try {
     const response = await fetch(url, {
       method: "DELETE",
       body: JSON.stringify(favoriteData),
-      headers: { "Content-Type": "application/json" }
+      headers: { 
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
     });
-    const favoriteId = await response.json();
-    return favoriteId;
+    return await response.json();
   } catch (error) {
     throw new Error("Error removing favorite");
   }
